refactor(graphql): split backward_consistency into backward_view modules
Separate consistent and historical query logic into distinct modules:
- backward_view::consistent — consistent view at a specific checkpoint
- backward_view::historical — exact id+version lookups without filtering
- backward_view::mod — shared constants and merge_and_deduplicate

Removes the BackwardView enum; callers use the module directly.
